third_party/WebKit/LayoutTests/imported/csswg-test/css-writing-modes-3/text-orientation-parsing-001-expected.txt

This is a testharness.js-based test.
PASS text-orientation initial value 
PASS -webkit-text-orientation: upright 
PASS -webkit-text-orientation: sideways 
PASS text-orientation should inherit 
FAIL -webkit-text-orientation: mixed assert_equals: expected "mixed" but got "upright"
PASS -webkit-text-orientation: initial 
PASS -webkit-text-orientation: inherit 
PASS -webkit-text-orientation: foo 
Harness: the test ran to completion.
